feat: 누적 리포트 페이지 컴포넌트 생성
